pyproject.toml
[tool.poetry]
name = "slack-secret"
version = "0.1.1"
description = ""
authors = ["Nicolas Paris <ni.paris@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tqdm = "^4.46.0"
slackclient = "^2.6.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
slacksecrets = 'slack_secret.main:cli'